requirements.txt
# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# LLM API clients
openai>=1.0.0
google-generativeai>=0.3.0

# Templating
jinja2>=3.1.0

# Numerical
numpy>=1.24.0

# Utilities
colorama>=0.4.6
httpx>=0.25.0

# Type checking (optional, for development)
# mypy>=1.0.0
# types-requests>=2.31.0
